Added defaults, expanded testing, added comments
Removed the dependencies from setup.py, cuz that's just a hassle

.travis.yml
- Switched from only testing `pyroclast/cpvae` to `pyroclast`

eager_run
- Removed `setup_data` fn and replaced with a call to
`common.tf_util.setup_tfds`
- Fixed `get_learn_function_defaults`
- Switched device logging to only happen in debugging

common.cmd_util
- Removed some unused command line arguments from `common_arg_parser`

common.tf_util
- Switched `setup_tfds` to use explicit args rather than a dictionary

cpvae
- Renamed `learn` to `cpvae` to follow the convention of naming each
submodule which contains the `learn` fn after the module
- Renamed `cpvae` to `model` to more clearly represent what it contains
and make space for previous renaming

Added cpvae.defaults to provide default args per dataset

cpvae.learn_test
- Added an explicit data_limit and output_dir. Now tests will output to
`./cpvae_learn_test/`

cpvae.model
- Added `self.output_fn` which is used in generation and set at
construction time by the `output_dist` arguement string. This should
move us closer to using actual distributions as the output.

cpvae.model_test
- Added an explicit data_limit and output_dir. Now tests will output to
`./cpvae_model_test/`

cpvae.tf_models
- Renamed `Encoder` to `VAEEncoder` and `Decoder` to `VAEDecoder`
- Added names to the decoder output layer so warnings of no gradients
are more informative

selfboosting
- Commented out sonnet imports and tests to not interfere with testing
the full library (it's all old code in there)
