+++ pyroclast/cpvae/__init__.py
@@ -1,2 +1,2 @@
-from pyroclast.cpvae.learn import learn
+from pyroclast.cpvae.cpvae import learn
 # from pyroclast.cpvae.generate import learn
