+++ pyroclast/cpvae/tf_models.py
@@ -3,10 +3,10 @@ import tensorflow as tf
 from pyroclast.common.models import get_network_builder
 
 
-class Encoder(tf.keras.Model):
+class VAEEncoder(tf.keras.Model):
 
     def __init__(self, network_name, latent_dim, name='enc'):
-        super(Encoder, self).__init__(name=name)
+        super(VAEEncoder, self).__init__(name=name)
         self.net = get_network_builder(network_name)()
         self.loc = tf.keras.layers.Dense(latent_dim, name='encoder_loc')
         self.scale = tf.keras.layers.Dense(latent_dim,
@@ -23,15 +23,15 @@ class Encoder(tf.keras.Model):
         return self.loc(embed), scale
 
 
-class Decoder(tf.keras.Model):
+class VAEDecoder(tf.keras.Model):
 
     def __init__(self, network_name, output_channels, name='dec'):
-        super(Decoder, self).__init__(name=name)
+        super(VAEDecoder, self).__init__(name=name)
         self.net = get_network_builder(network_name)()
-        self.loc = tf.keras.layers.Conv2D(output_channels, 3, padding="same")
+        self.loc = tf.keras.layers.Conv2D(output_channels, 3, padding="same", name='decoder_loc')
         self.log_scale = tf.keras.layers.Conv2D(output_channels,
                                                 3,
-                                                padding="same")
+                                                padding="same", name='decoder_log_scale')
 
     def call(self, z):
         latent = self.net(z)
