+++ pyroclast/eager_run.py
@@ -4,13 +4,11 @@ import sys
 from importlib import import_module
 
 import tensorflow as tf
-import tensorflow_datasets as tfds
 from tensorflow.python.client import device_lib
 
 from pyroclast.common.cmd_util import common_arg_parser, parse_unknown_args
-from pyroclast.common.util import img_preprocess
+from pyroclast.common.tf_util import setup_tfds
 
-print(device_lib.list_local_devices())
 tf.compat.v1.enable_eager_execution()
 
 
@@ -36,10 +34,8 @@ def get_alg_module(alg, submodule=None):
     try:
         # first try to import the alg module from baselines
         alg_module = import_module('.'.join(['pyroclast',
-                                             alg]))  #, submodule]))
+                                             alg, submodule]))
     except ImportError:
-        # then from rl_algs
-        # alg_module = import_module('.'.join(['rl_' + 'algs', alg, submodule]))
         print('failed to import from {}'.format('.'.join(['pyroclast', alg])))
         assert False
 
@@ -50,66 +46,28 @@ def get_learn_function(alg):
     return get_alg_module(alg).learn
 
 
-def get_learn_function_defaults(alg, env_type):
+def get_learn_function_defaults(alg, dataset):
     try:
         alg_defaults = get_alg_module(alg, 'defaults')
-        kwargs = getattr(alg_defaults, env_type)()
+        kwargs = getattr(alg_defaults, dataset)()
     except (ImportError, AttributeError):
         kwargs = {}
     return kwargs
 
 
-def setup_data(args):
-    # load data
-    data_dict, info = tfds.load(args.dataset,
-                                with_info=True,
-                                data_dir='./data/')
-    if args.dataset == 'celeb_a':
-        data_dict['num_classes'] = 2
-        data_dict['train_bpe'] = info.splits['train'].num_examples // args[
-            'batch_size']
-        data_dict[
-            'test_bpe'] = info.splits['test'].num_examples // args['batch_size']
-        data_dict['shape'] = info.features['image'].shape
-
-        data_dict['all_train'] = data_dict['train']
-        data_dict['train'] = data_dict['train'].map(
-            lambda x: img_preprocess(x, args.image_size)).shuffle(1024).batch(
-                args['batch_size'])
-        data_dict['all_test'] = data_dict['test']
-        data_dict['test'] = data_dict['test'].map(
-            lambda x: img_preprocess(x, args.image_size)).batch(
-                args['batch_size'])
-    else:
-        data_dict[
-            'train_bpe'] = info.splits['train'].num_examples // args.batch_size
-        data_dict[
-            'test_bpe'] = info.splits['test'].num_examples // args.batch_size
-        data_dict['shape'] = info.features['image'].shape
-        data_dict['num_classes'] = info.features['label'].num_classes
-
-        data_dict['train'] = data_dict['train'].shuffle(1024).batch(
-            args.batch_size).prefetch(tf.data.experimental.AUTOTUNE)
-        data_dict['test'] = data_dict['test'].batch(args.batch_size).prefetch(
-            tf.data.experimental.AUTOTUNE)
-
-    return data_dict
-
-
 def train(args, extra_args):
     # load data
-    seed = args.seed
-
     learn = get_learn_function(args.alg)
     alg_kwargs = get_learn_function_defaults(args.alg, args.dataset)
     alg_kwargs.update(extra_args)
-    data_dict = setup_data(args)
+    data_dict = setup_tfds(args.dataset, args.batch_size, args.data_limit, args.data_dir)
 
     print('Training {} on {} with arguments \n{}'.format(
         args.alg, args.dataset, alg_kwargs))
     model = learn(data_dict,
-                  seed=seed,
+                  seed=args.seed,
                   output_dir=args.output_dir,
+                  debug=args.debug,
                   **alg_kwargs)
 
     return model
@@ -119,6 +77,9 @@ def main(args):
     arg_parser = common_arg_parser()
     args, unknown_args = arg_parser.parse_known_args(args)
     extra_args = parse_cmdline_kwargs(unknown_args)
+
+    if args.debug:
+        print(device_lib.list_local_devices())
 
     # save the parameters
     if not os.path.exists(args.output_dir):
