+++ pyroclast/selfboosting/sequential_resnet.py
@@ -1,4 +1,4 @@
-import sonnet as snt
+#import sonnet as snt
 import tensorflow as tf
 
 from pyroclast.selfboosting.residual_boosting_module import ResidualBoostingModule
